Altered the display of SuperTransformation's ability name in the set bonus.
Added a wish to receive individual chaos emeralds.

Altered the wish that gives you all the chaos emeralds to loop over the method wish that gives you one of them.

restored some of the older HP values for prickle pigs. They get aded on top of their sValue.

Added a flag to hopefully prevent triggering a second blink whilst in the process of completing a first one.